REFACTOR : 컴포넌트별 불필요한 함수 이동
