+++ 0617/src/components/Landing/ReviewForm.jsx
@@ -3,20 +3,106 @@ import styles from './reviewForm.module.scss';
 import { MdOutlineCameraAlt } from 'react-icons/md';
 import { FaTimes } from 'react-icons/fa';
 import Spinner from '@/components/common/Spinner/index';
+import { useRef, useState } from 'react';
+import supabase from '@/apis/supabaseApi';
 
 const cx = classNames.bind(styles);
 
-function ReviewForm({
-  user,
-  handleInput,
-  handleUploadAndSubmit,
-  textareaRef,
-  previewUrl,
-  handleFileChange,
-  handleCancelSelection,
-  uploadImg,
-}) {
-  console.log(uploadImg);
+function ReviewForm({ user, productId, reviewList, setReviewList }) {
+  const textareaRef = useRef(null);
+  const [file, setFile] = useState(null);
+  const [previewUrl, setPreviewUrl] = useState(null);
+  const [uploadImg, setUploading] = useState(false);
+
+  const handleInput = (event) => {
+    const { value } = event.target;
+    if (value.length <= 3000) {
+      adjustTextareaHeight();
+    }
+  };
+
+  const adjustTextareaHeight = () => {
+    if (textareaRef.current) {
+      const minHeight = 50;
+      textareaRef.current.style.height = 'auto';
+      const scrollHeight = textareaRef.current.scrollHeight;
+      if (scrollHeight >= minHeight) {
+        textareaRef.current.style.height = `${scrollHeight + 4}px`;
+      }
+    }
+  };
+
+  const handleFileChange = (event) => {
+    const selectedFile = event.target.files[0];
+    setFile(selectedFile);
+    setPreviewUrl(URL.createObjectURL(selectedFile));
+  };
+
+  const handleUploadAndSubmit = async () => {
+    try {
+      setUploading(true);
+      let imageUrl = null;
+      if (textareaRef.current.value === '') {
+        alert('내용을 입력해주세요.');
+        setUploading(false);
+        return;
+      }
+
+      if (file) {
+        const fileExt = file.name.split('.').pop();
+        const fileName = `${Date.now()}.${fileExt}`;
+        const filePath = `public/${fileName}`;
+
+        let { error: uploadError } = await supabase.storage
+          .from('images')
+          .upload(filePath, file, {
+            cacheControl: '3600',
+            upsert: false,
+          });
+
+        if (uploadError) {
+          throw uploadError;
+        }
+
+        const publicUrl = supabase.storage
+          .from('images')
+          .getPublicUrl(filePath);
+
+        imageUrl = publicUrl.data.publicUrl;
+        setPreviewUrl(null);
+        setFile(null);
+      }
+
+      const { data, error } = await supabase
+        .from('reviews')
+        .insert([
+          {
+            product_id: productId,
+            username: user.email,
+            review_text: textareaRef.current.value.trim(),
+            review_img: imageUrl,
+          },
+        ])
+        .select();
+      if (error) {
+        return;
+      } else {
+        setReviewList([...reviewList, ...data]);
+        textareaRef.current.value = '';
+      }
+    } catch (error) {
+      console.log(error);
+      alert('Error uploading image and submitting review!');
+    } finally {
+      setUploading(false);
+    }
+  };
+
+  const handleCancelSelection = () => {
+    setPreviewUrl(null);
+    setFile(null);
+  };
+
   return (
     <article className={cx('reviewBox')}>
       <div className={cx('reviewHeader')}>
